+++ src/modules/person-kontext/api/personenkontext.uc.ts
@@ -4,8 +4,8 @@ import { Inject, Injectable } from '@nestjs/common';
 import { Paged } from '../../../shared/paging/paged.js';
 import { PersonDo } from '../../person/domain/person.do.js';
 import { PersonService } from '../../person/domain/person.service.js';
-import { PersonenkontextDo } from '../../person/domain/personenkontext.do.js';
-import { PersonenkontextService } from '../../person/domain/personenkontext.service.js';
+import { PersonenkontextDo } from '../domain/personenkontext.do.js';
+import { PersonenkontextService } from '../domain/personenkontext.service.js';
 import { CreatePersonenkontextDto } from './create-personenkontext.dto.js';
 import { CreatedPersonenkontextDto } from './created-personenkontext.dto.js';
 import { FindPersonenkontextByIdDto } from './find-personenkontext-by-id.dto.js';
