+++ src/modules/person-kontext/domain/personenkontext.enums.ts
