+++ src/modules/person-kontext/domain/personenkontext.service.spec.ts
@@ -7,9 +7,9 @@ import { DomainError } from '../../../shared/error/domain.error.js';
 import { EntityCouldNotBeCreated } from '../../../shared/error/entity-could-not-be-created.error.js';
 import { EntityNotFoundError } from '../../../shared/error/entity-not-found.error.js';
 import { Paged } from '../../../shared/paging/paged.js';
-import { PersonRepo } from '../persistence/person.repo.js';
-import { PersonenkontextRepo } from '../persistence/personenkontext.repo.js';
-import { PersonDo } from './person.do.js';
+import { PersonRepo } from '../../person/persistence/person.repo.js';
+import { PersonenkontextRepo } from '../../person/persistence/personenkontext.repo.js';
+import { PersonDo } from '../../person/domain/person.do.js';
 import { PersonenkontextDo } from './personenkontext.do.js';
 import { PersonenkontextService } from './personenkontext.service.js';
 import { MismatchedRevisionError } from '../../../shared/error/mismatched-revision.error.js';
