+++ src/modules/person/api/person.uc.spec.ts
@@ -8,8 +8,8 @@ import { Paged } from '../../../shared/paging/index.js';
 import { KeycloakUserService } from '../../keycloak-administration/index.js';
 import { PersonDo } from '../domain/person.do.js';
 import { PersonService } from '../domain/person.service.js';
-import { SichtfreigabeType } from '../domain/personenkontext.enums.js';
-import { PersonenkontextService } from '../domain/personenkontext.service.js';
+import { SichtfreigabeType } from '../../person-kontext/domain/personenkontext.enums.js';
+import { PersonenkontextService } from '../../person-kontext/domain/personenkontext.service.js';
 import { CreatePersonDto } from './create-person.dto.js';
 import { FindPersonendatensatzDto } from './find-personendatensatz.dto.js';
 import { PersonApiMapperProfile } from './person-api.mapper.profile.js';
