+++ src/modules/person/api/person.uc.ts
@@ -8,9 +8,9 @@ import { Paged } from '../../../shared/paging/index.js';
 import { KeycloakUserService } from '../../keycloak-administration/index.js';
 import { PersonDo } from '../domain/person.do.js';
 import { PersonService } from '../domain/person.service.js';
-import { PersonenkontextDo } from '../domain/personenkontext.do.js';
-import { SichtfreigabeType } from '../domain/personenkontext.enums.js';
-import { PersonenkontextService } from '../domain/personenkontext.service.js';
+import { PersonenkontextDo } from '../../person-kontext/domain/personenkontext.do.js';
+import { SichtfreigabeType } from '../../person-kontext/domain/personenkontext.enums.js';
+import { PersonenkontextService } from '../../person-kontext/domain/personenkontext.service.js';
 import { CreatePersonDto } from './create-person.dto.js';
 import { FindPersonendatensatzDto } from './find-personendatensatz.dto.js';
 import { FindPersonenkontextDto } from '../../person-kontext/api/find-personenkontext.dto.js';
