+++ src/modules/person/persistence/person-persistence.mapper.profile.ts
@@ -2,7 +2,7 @@ import { Mapper, MappingProfile, createMap, forMember, ignore } from '@automappe
 import { AutomapperProfile, getMapperToken } from '@automapper/nestjs';
 import { Inject, Injectable } from '@nestjs/common';
 import { PersonDo } from '../domain/person.do.js';
-import { PersonenkontextDo } from '../domain/personenkontext.do.js';
+import { PersonenkontextDo } from '../../person-kontext/domain/personenkontext.do.js';
 import { PersonEntity } from '../persistence/person.entity.js';
 import { PersonenkontextEntity } from './personenkontext.entity.js';
