+++ src/modules/person/persistence/personenkontext.repo.integration-spec.ts
@@ -11,8 +11,8 @@ import {
     MapperTestModule,
 } from '../../../../test/utils/index.js';
 import { PersonDo } from '../domain/person.do.js';
-import { PersonenkontextDo } from '../domain/personenkontext.do.js';
-import { Personenstatus, Rolle, SichtfreigabeType } from '../domain/personenkontext.enums.js';
+import { PersonenkontextDo } from '../../person-kontext/domain/personenkontext.do.js';
+import { Personenstatus, Rolle, SichtfreigabeType } from '../../person-kontext/domain/personenkontext.enums.js';
 import { PersonPersistenceMapperProfile } from './person-persistence.mapper.profile.js';
 import { PersonEntity } from './person.entity.js';
 import { PersonenkontextEntity } from './personenkontext.entity.js';
