+++ src/modules/person/person.module.spec.ts
@@ -4,7 +4,7 @@ import { PersonService } from './domain/person.service.js';
 import { PersonRepo } from './persistence/person.repo.js';
 import { PersonModule } from './person.module.js';
 import { PersonPersistenceMapperProfile } from './persistence/person-persistence.mapper.profile.js';
-import { PersonenkontextService } from './domain/personenkontext.service.js';
+import { PersonenkontextService } from '../person-kontext/domain/personenkontext.service.js';
 import { PersonenkontextRepo } from './persistence/personenkontext.repo.js';
 
 describe('PersonModule', () => {
